Split up world module
